feat(gwasUnifiedFlowReacTour): Initial working version of ReacTour for unified flow
