+++ src/Analysis/AnalysisApp.jsx
@@ -20,6 +20,14 @@ const queryClient = new QueryClient();
 const disableBody = (target) => disableBodyScroll(target);
 const enableBody = (target) => enableBodyScroll(target);
 
+const steps = [
+  {
+  selector: '.first-step',
+  content: 'This is my first Step',
+  }
+];
+
+
 class AnalysisApp extends React.Component {
   constructor(props) {
     super(props);
@@ -121,9 +129,11 @@ class AnalysisApp extends React.Component {
       );
     case 'GWAS++': {
       return (
+        <TourProvider steps={steps}>
         <div>
           <GWASContainer refreshWorkflows={this.refreshWorkflows} />
         </div>
+        </TourProvider>
       );
     }
     default:
